backend-django/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.6
redis==5.2.1
celery==5.4.0
